refactor(kamn-node): extract scope fixture projection parser submodule
